util/auto/krebs: initial commit
the auto/krebs utility is a tool to easily deploy the krebscode/painload
on every machine.
it will utilize the punani packet manager of krebs in order to install
all the packets you want e.g. retiolum to join the 42::/16 network.
The packet manager will use god mode in order to work correctly.