gold/scoreboard: initial commit
Add a copy/paste branch of the original btc_logger. I couldn't import
the whole repo as it contains tons of passwords and keys (which is not a
good thing as these are somewhat special for me :)

scoreboard contains a number of modules which do a number of things:
ati_temper: log temperature data of an ati_temper enabled host :
  see //gold/miner/bin/ati_temper
solidcoin : graph solidcoin informations
bitcoin   : s.a.

In addition to that live stats are generated by the script

//gold/scoreboard/profile contains all the important paths and variables
which are sourced by the modules. In a new installation this file needs
to be customized.

//gold/scoreboard/util/schedule_logger is the scheduler for generating
the graphs every 5 minutes.

you will need to run /modules/available/$MODULE/init before using the
scoreboard as a whole.

the modules directory is still a bit cluttered, sorry about that.